Made the PollingObserver also find files already existing
When starting the PollingObserver the files and directories already
existing in the path to watch will create an 'existing' event to indicate
that files/dirs are already existing. This will in the end call the method
on_existing.

This can be useful if you want to watch a path for changes, but are
interested in the files already located there to start with. The benefit
of not creating a custom solution around the Watchdog is that the event
handlers run in their own threads and we can reuse the same mechanism for
retrieving the events